pyproject.toml
[project]
name = "gemini-research-mcp"
dynamic = ["version"]
description = "MCP server for AI-powered research using Gemini: quick grounded search + Deep Research Agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Guillaume Fortaine", email = "guillaume.fortaine+github@machinemates.ai" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "gemini",
    "research",
    "ai",
    "deep-research",
    "grounding",
    "google-search",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Typing :: Typed",
]
dependencies = [
    "google-genai>=1.59.0",
    # MCP upgraded from 1.10.0 to 1.25.0; restrict to 1.x to avoid unexpected breaking changes.
    "mcp>=1.25.0,<2.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    # Persistent storage using industry-standard key-value interface (used by FastMCP, agentpool, etc.)
    "py-key-value-aio[disk]>=0.3.0",
    # XDG-compliant platform-specific directories
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
# DOCX export support with TOC, hyperlinks, bookmarks (skelmis-docx is an enhanced python-docx fork)
# Using fork until upstream merges pywin32 fix: https://github.com/Skelmis/python-docx/pull/30
# Pinned to commit SHA for reproducibility
docx = [
    "skelmis-docx @ git+https://github.com/fortaine/python-docx.git@06d0809bb29f561051dfcc142c03a9d96f439f89",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.3.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    # Using fork until upstream merges pywin32 fix: https://github.com/Skelmis/python-docx/pull/30
    # Pinned to commit SHA for reproducibility
    "skelmis-docx @ git+https://github.com/fortaine/python-docx.git@fix/pywin32-version-constraint",
]

[project.urls]
Homepage = "https://github.com/fortaine/gemini-research-mcp"
Documentation = "https://github.com/fortaine/gemini-research-mcp#readme"
Repository = "https://github.com/fortaine/gemini-research-mcp"
Issues = "https://github.com/fortaine/gemini-research-mcp/issues"

[project.scripts]
gemini-research-mcp = "gemini_research_mcp.server:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.metadata]
# Allow git+ references for dependencies (needed for fork while waiting for upstream PR merge)
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_research_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests that require GEMINI_API_KEY",
]
addopts = "-m 'not e2e'"

[tool.mypy]
strict = true
python_version = "3.12"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
